move queue configuration to config class
