+++ src/main/java/uk/gov/ons/fwmt/job_service_v2/ApplicationConfig.java
@@ -41,63 +41,4 @@ public class ApplicationConfig {
     SpringApplication.run(ApplicationConfig.class, args);
     log.debug("Started application");
   }
-
-  @Bean
-  Queue adapterQueue() {
-    return new Queue(QueueConfig.JOBSVC_TO_ADAPTER_QUEUE, false);
-  }
-
-  @Bean
-  Queue jobsvcQueue() {
-    return new Queue(QueueConfig.ADAPTER_TO_JOBSVC_QUEUE, false);
-  }
-
-  @Bean
-  TopicExchange adapterExchange() {
-    return new TopicExchange(QueueConfig.RM_JOB_SVC_EXCHANGE);
-  }
-
-  @Bean
-  Binding adapterBinding(@Qualifier("adapterQueue") Queue queue, TopicExchange exchange) {
-    return BindingBuilder.bind(queue).to(exchange).with(QueueConfig.JOB_SVC_RESPONSE_ROUTING_KEY);
-  }
-
-  @Bean
-  Binding jobsvcBinding(@Qualifier("jobsvcQueue") Queue queue, TopicExchange exchange) {
-    return BindingBuilder.bind(queue).to(exchange).with(QueueConfig.JOB_SVC_REQUEST_ROUTING_KEY);
-  }
-
-  @Bean
-  SimpleMessageListenerContainer container(ConnectionFactory connectionFactory,
-      MessageListenerAdapter listenerAdapter) {
-    SimpleMessageListenerContainer container = new SimpleMessageListenerContainer();
-    container.setConnectionFactory(connectionFactory);
-    container.setQueueNames(QueueConfig.ADAPTER_TO_JOBSVC_QUEUE);
-    container.setMessageListener(listenerAdapter);
-    return container;
-  }
-
-  @Bean
-  MessageListenerAdapter listenerAdapter(JobServiceMessageReceiver receiver) {
-    return new MessageListenerAdapter(receiver, "receiveMessage");
-  }
-
-  @Bean
-  public RetryTemplate retryTemplate() {
-    RetryTemplate retryTemplate = new RetryTemplate();
-
-    ExponentialBackOffPolicy backOffPolicy = new ExponentialBackOffPolicy();
-    backOffPolicy.setInitialInterval(5000);
-    backOffPolicy.setMultiplier(3.0);
-    backOffPolicy.setMaxInterval(45000);
-    retryTemplate.setBackOffPolicy(backOffPolicy);
-
-    SimpleRetryPolicy retryPolicy = new SimpleRetryPolicy(3);
-    retryPolicy.setMaxAttempts(3);
-    retryTemplate.setRetryPolicy(retryPolicy);
-
-    retryTemplate.registerListener(new DefaultListenerSupport());
-
-    return retryTemplate;
-  }
 }
